refactor tweaks
