+++ src/frontend/src/types/zustand/tweaks/index.ts
@@ -1,8 +1,6 @@
 import { tweakType } from "../../components";
 
 export type TweaksStoreType = {
-  tweaksObject: tweakType;
-  setTweaksObject: (tweak: tweakType) => void;
   tweak: tweakType;
   setTweak: (tweak: tweakType) => void;
   tweaksList: string[];
